requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
clang==5.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.1.0
idna==3.10
importlib_metadata==8.5.0
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.7
MarkupSafe==2.1.5
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
protobuf==3.20.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.15.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-gpu==2.6.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.12.1
zipp==3.20.2
